Give each component their own file
