+++ packages/react-arborist/src/components/list.tsx
@@ -1,0 +1,26 @@
+import { FixedSizeList } from "react-window";
+import { useStaticContext } from "../context";
+import { ListOuterElement } from "./list-outer-element";
+import { Row } from "./row";
+
+export function List(props: { className?: string }) {
+  const tree = useStaticContext();
+  return (
+    <div style={{ height: tree.height, width: tree.width, overflow: "hidden" }}>
+      <FixedSizeList
+        className={props.className}
+        outerRef={tree.listEl}
+        itemCount={tree.api.visibleNodes.length}
+        height={tree.height}
+        width={tree.width}
+        itemSize={tree.rowHeight}
+        itemKey={(index) => tree.api.visibleNodes[index]?.id || index}
+        outerElementType={ListOuterElement}
+        // @ts-ignore
+        ref={tree.list}
+      >
+        {Row}
+      </FixedSizeList>
+    </div>
+  );
+}
