+++ packages/react-arborist/src/components/outer-drop.ts
@@ -1,0 +1,7 @@
+import { ReactElement } from "react";
+import { useOuterDrop } from "../dnd/outer-drop-hook";
+
+export function OuterDrop(props: { children: ReactElement }) {
+  useOuterDrop();
+  return props.children;
+}
